feat: change generateAddressAndSecret back to sync
Instead of making `Server#generateAddressAndSecret` async to accommodate the browser build (due to WebCrypto), don't include the `Server` in the browser build at all. With this change, the `dj-web-crypto` branch is no longer an API-breaking change.
