+++ test/connection.test.ts
@@ -22,7 +22,7 @@ describe('Connection', function () {
       serverSecret: Buffer.alloc(32)
     })
 
-    const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+    const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
     this.destinationAccount = destinationAccount
     this.sharedSecret = sharedSecret
 
@@ -484,7 +484,7 @@ describe('Connection', function () {
         toFake: ['setTimeout', 'Date']
       })
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
       const errPromise = new Promise((resolve, reject) => {
@@ -564,7 +564,7 @@ describe('Connection', function () {
       this.clientPlugin.exchangeRate = 0.0000001
       this.clientPlugin.maxAmount = 1000000000 // 10^9
       await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
     })
@@ -573,7 +573,7 @@ describe('Connection', function () {
       this.clientPlugin.exchangeRate = 0.0000000001
       this.clientPlugin.maxAmount = 1000000000000 // 10^12
       await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
     })
@@ -582,7 +582,7 @@ describe('Connection', function () {
       this.clientPlugin.exchangeRate = 0.0000000001
       this.serverPlugin.exchangeRate = 1 / this.clientPlugin.exchangeRate
       await assert.isRejected(createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       }), 'Error connecting: Unable to establish connection, no packets meeting the minimum exchange precision of 3 digits made it through the path.')
     })
@@ -590,7 +590,7 @@ describe('Connection', function () {
     it('should apply a default slippage of 1% to the exchange rate', async function () {
       const slippage = 0.01
       const connection = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
       const exchangeRateWithSlippage = (0.5 * (1 - slippage)).toString()
@@ -600,7 +600,7 @@ describe('Connection', function () {
     it('should apply slippage to the exchange rate when explicitly specified', async function () {
       const slippage = 0.05
       const connection = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         slippage
       })
@@ -618,7 +618,7 @@ describe('Connection', function () {
       this.clientPlugin.maxAmount = 150000
 
       const connection = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         slippage: 0 
       })
@@ -638,7 +638,7 @@ describe('Connection', function () {
       this.clientPlugin.maxAmount = 1500
 
       await assert.isRejected(createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       }), 'Error connecting: Unable to establish connection, no packets meeting the minimum exchange precision of 3 digits made it through the path.')
       clearInterval(interval)
@@ -668,7 +668,7 @@ describe('Connection', function () {
         }))
 
       await assert.isRejected(createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       }), 'Error connecting: Unable to establish connection, no packets meeting the minimum exchange precision of 3 digits made it through the path.')
 
@@ -704,7 +704,7 @@ describe('Connection', function () {
         }))
         .callThrough()
       await assert.isRejected(createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       }), 'Error connecting: Unable to establish connection, no packets meeting the minimum exchange precision of 3 digits made it through the path.')
       clearInterval(interval)
@@ -739,7 +739,7 @@ describe('Connection', function () {
         }))
         .callThrough()
       await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         minExchangeRatePrecision: 1,
         plugin: this.clientPlugin
       })
@@ -812,7 +812,7 @@ describe('Connection', function () {
         .callsFake(mySendData)
 
       await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         minExchangeRatePrecision: 2,
         plugin: this.clientPlugin
       })
@@ -844,7 +844,7 @@ describe('Connection', function () {
 
       const serverPromise = this.server.acceptConnection()
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
 
@@ -875,7 +875,7 @@ describe('Connection', function () {
         }))
 
       await assert.isRejected(createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       }), 'Error connecting: Unable to establish connection, no packets meeting the minimum exchange precision of 3 digits made it through the path.')
 
@@ -925,7 +925,7 @@ describe('Connection', function () {
       })
       await this.server.listen()
 
-      const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+      const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
       this.destinationAccount = destinationAccount
       this.sharedSecret = sharedSecret
 
@@ -981,7 +981,7 @@ describe('Connection', function () {
       })
       await this.server.listen()
 
-      const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+      const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
       this.destinationAccount = destinationAccount
       this.sharedSecret = sharedSecret
 
@@ -1040,7 +1040,7 @@ describe('Connection', function () {
     it('should find the maximum amount immediately if the connector returns the receivedAmount and maximumAmount in the F08 error data', async function () {
       this.clientPlugin.maxAmount = 1500
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
       assert.equal(clientConn['maximumPacketAmount'].toString(), '1500')
@@ -1147,7 +1147,7 @@ describe('Connection', function () {
 
       const serverPromise = this.server.acceptConnection()
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin
       })
       const serverConn = await serverPromise
@@ -1358,7 +1358,7 @@ describe('Connection', function () {
         enablePadding: true
       })
 
-      const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+      const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
       this.destinationAccount = destinationAccount
       this.sharedSecret = sharedSecret
 
@@ -1398,7 +1398,7 @@ describe('Connection', function () {
 
   describe('Stream IDs', function () {
     it('should close the connection if the peer uses the wrong numbered stream ID', async function () {
-      const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+      const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
       const clientPlugin = this.clientPlugin
       class BadConnection extends Connection {
         constructor() {
@@ -1432,7 +1432,7 @@ describe('Connection', function () {
 
     it('should close the connection if the peer opens too many streams', async function () {
       const spy = sinon.spy()
-      const { destinationAccount, sharedSecret } = await this.server.generateAddressAndSecret()
+      const { destinationAccount, sharedSecret } = this.server.generateAddressAndSecret()
       this.server.on('connection', (serverConn: Connection) => {
         serverConn['maxStreamId'] = 6
         serverConn.on('stream', (stream: DataAndMoneyStream) => {
@@ -1476,7 +1476,7 @@ describe('Connection', function () {
     it('should allow the user to set the maximum number of open streams', async function () {
       const serverConnPromise = this.server.acceptConnection()
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         maxRemoteStreams: 1
       })
@@ -1500,7 +1500,7 @@ describe('Connection', function () {
     it('should throw an error when the user calls createStream if it would exceed the other side\'s limit', async function () {
       const serverConnPromise = this.server.acceptConnection()
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         maxRemoteStreams: 2
       })
@@ -1518,7 +1518,7 @@ describe('Connection', function () {
     it('should increase the max stream id as streams are closed', async function () {
       const serverConnPromise = this.server.acceptConnection()
       const clientConn = await createConnection({
-        ...await this.server.generateAddressAndSecret(),
+        ...this.server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         maxRemoteStreams: 2
       })
@@ -1619,7 +1619,7 @@ describe('Connection', function () {
       const serverConnPromise = server.acceptConnection()
 
       const clientConn = await createConnection({
-        ...await server.generateAddressAndSecret(),
+        ...server.generateAddressAndSecret(),
         plugin: this.clientPlugin,
         connectionBufferSize: 2500
       })
